+++ src/index.ts
@@ -3,3 +3,4 @@ import "reflect-metadata";
 export * from "./koa";
 export * from "./openapi";
 export * from "./references";
+export * from "./special";