+++ src/openapi/component-schema.ts
@@ -1,4 +1,5 @@
 import { Constructor } from "../common/declares";
+import { DisplayName, getDisplayName } from "../special/display-name";
 import { checkConstructorProperty } from "../common/functions";
 import { toJSONSchema } from "./functions";
 
@@ -13,7 +14,7 @@ export const componentsSet = new Set();
 export function getComponentsSchemas(): Record<string, Constructor> {
   const result = {};
   componentsSet.forEach((constructor: Constructor) => {
-    const { name } = constructor;
+    const name = getDisplayName(constructor);
     Object.assign(result, { [name]: toJSONSchema(constructor) });
   });
   return result;
@@ -23,9 +24,14 @@ export function getComponentsSchemas(): Record<string, Constructor> {
  * декоратор для записи класса в структуру определений
  * @returns {ClassDecorator}
  */
-export function ComponentSchema(): ClassDecorator {
+export function ComponentSchema(displayName?: string): ClassDecorator {
   return (constructor) => {
     checkConstructorProperty(constructor);
+    // если указано видимое имя, то добавим его для элемента данных
+    if (displayName) {
+      Reflect.decorate([DisplayName(displayName)], constructor);
+    }
+    // сохраним в справочнике
     componentsSet.add(constructor);
   };
 }