+++ src/openapi/functions.ts
@@ -3,8 +3,10 @@ import { SchemaObject } from "openapi3-ts";
 import { defaultMetadataStorage } from "class-transformer/cjs/storage";
 import { refPointerPrefix } from "./component-schema";
 
+const noJSONSchemaSet = new Set();
+
 export function toJSONSchema(constructor): SchemaObject {
-  if (typeof constructor === "function") {
+  if (typeof constructor === "function" && !noJSONSchemaSet.has(constructor)) {
     return targetConstructorToSchema(constructor, {
       classTransformerMetadataStorage: defaultMetadataStorage,
       refPointerPrefix,
@@ -12,4 +14,10 @@ export function toJSONSchema(constructor): SchemaObject {
   } else {
     return <SchemaObject>constructor;
   }
+}
+
+export function NoJSONSchema(): ClassDecorator {
+  return (constructor) => {
+    noJSONSchemaSet.add(constructor);
+  };
 }